+++ frontend/src/themes/vs-dark-plus-theme.ts
