+++ frontend/src/utils/configuration.ts
